Fixed the list of bleeding codes to use the rhic library [note]
The previous version of the codes in the R script contained an error
that I hadn't spotted -- the R processing was performing exact matches
against the codes list, but the list contained items like I60 which were
actually categories (containing many codes). As a result, many codes were
being missed out of the analysis. A way to fix this would have been to use
a wildcard-type regex in the match (i.e. strip whitespace, convert to lower
case, and then compare against i60* for inclusion). Then full codes like
I21.0 would work too, because the code would be compared against
i210*, which would only match i210 but not i212 (say). I believe there are
no instances where this would not work in this case.