Found errors in the HES spells dataset
There are way too many index events, which points to
an issue with how spells are being identified as an
index event. It looks like there are approximately the
right number of raw spell rows (assuming that on
average spells contain one episode). The index event
count is wrong all the way from idx_spells on line 244.

That variable is calculated as mi_schnier_count > 0 or
pci_count > 0. This implies that one or other of those
groups is wrong, or is different from the previous
analysis in a way that admits many more index events.